third_party/WebKit/LayoutTests/fast/css/atapply/at-apply-cascade.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
#e {
    --foo: {
        background-color: green;
        width: 10px;
        height: 200px;
    };
    background-color: red;
    @apply --foo;
    width: 100px;
}
</style>
<div id=e>
</div>
<script>
test(function(){
    assert_equals(getComputedStyle(e).backgroundColor, "rgb(0, 128, 0)");
    assert_equals(getComputedStyle(e).width, "100px");
    assert_equals(getComputedStyle(e).height, "200px");
}, "Declarations from @apply cascade correctly within a single rule");
</script>
